requirements.txt
django
djangorestframework
scrapy
BeautifulSoup4